8. blogs_platform/requirements.txt
Django==2.1.4
django-hitcount==1.3.0
pytz==2018.7
